test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pytest>=2.6.4
coverage>=3.6
mock>=1.0.1
docker-py>=1.1.0
path.py>=5.2
